COLLECTION_SCOPES as a hash w/ scopes' arrays
Problem:

As the number of scopes defined in the COLLECTION_SCOPES array grows they start
to fill all the available space in the index header. Besides, that structure
don't let us know which scopes can be mixed in order to refine our search.

Solution:

Let the user define COLLECTION_SCOPES as a hash of scopes's arrays in which
each key represents a different concept (e.g. color, size, price...). Each key
can be used in the index header to arrange all the scope buttons.