"Disable scopes" feature and used by default (in the generated dashboard)
Problem
Use of scopes while searching assumes that the dashboard has been properly
secured. In some scenarios scopes can be very dangerous and currently there
isn't a way to disabled them in a dashboard.

Solution
The definition of COLLECTION_SCOPES as an empty hash/array will disable scopes
for that dashboard. When generating a new dashboard that will be the default
configuration.